+++ CHANGELOG.md
@@ -9,6 +9,9 @@
 - Add `garn.callFlake` helper to allow importing flake files by local path.
 - Add `garn.importFlake` helper to allow importing flake files by url, e.g. from GitHub.
 
+- Add `Environment.addToSetup` and `Environment.addToSandboxSetup`, making setups for `Environment`s more flexible.
+- Remove `sandboxSetup` parameter from `mkEnvironment`, use `addtoSandboxSetup` instead.
+
 ## v0.0.18
 
 - Allow entering top-level `Environment`s with `garn enter`.